--- packages/badge/css.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/cjs/css')
