+++ packages/badge/package.json
@@ -6,19 +6,20 @@
   "repository": "pluralsight/design-system",
   "author": "jaketrent",
   "keywords": [],
-  "main": "dist/cjs/index.js",
+  "main": "dist/esm/index.js",
   "module": "dist/esm/index.js",
   "scripts": {
-    "build": "run-s build:cjs build:esm build:css",
-    "build:cjs": "tsc --project tsconfig.build.json --module commonjs --target es5 --outDir dist/cjs",
-    "build:css": "build-css --useGlamor -i dist/cjs/css/index.js",
+    "build": "run-s build:esm build:css",
+    "build:css": "build-css --useGlamor -i dist/esm/css/index.js",
     "build:esm": "tsc --project tsconfig.build.json --module es2015 --target es5 --outDir dist/esm",
     "build:watch": "yarn build:js --watch",
     "storybook": "start-storybook -p 6006",
     "test": "jest",
     "test:watch": "yarn test --watchAll"
   },
+  "sideEffects": false,
   "style": "dist/css/index.css",
+  "type": "module",
   "dependencies": {
     "@pluralsight/ps-design-system-core": "^6.9.3",
     "@pluralsight/ps-design-system-util": "^7.2.1"
