+++ packages/badge/src/js/index.ts
@@ -1,7 +1,7 @@
 import Theme from '@pluralsight/ps-design-system-theme'
 import { ValueOf } from '@pluralsight/ps-design-system-util'
 
-import * as vars from '../vars'
+import * as vars from '../vars/index.js'
 
 export function select(
   themeName: ValueOf<typeof Theme.names>,
